Refactor type declarations for lucide-react and sonner modules
