+++ tsconfig.json
@@ -27,6 +27,10 @@
         "name": "next"
       }
     ],
+    "typeRoots": [
+      "types",
+      "node_modules/@types"
+    ],
     "types": [
       "node",
       "react",
